[Merge] From BRANCH_1_3_STABLE, r2265: Eliminated compilation warnings under Visual C++ 2005.
git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@2950 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ orig/trunk/vnc_winsrc/vncviewer/FileTransfer.cpp
@@ -185,7 +185,7 @@ FileTransfer::uploadFilePortion()
 				PostMessage(m_pCC->m_hwnd, (UINT) WM_FT_CHECKTRANSFERQUEUE, (WPARAM) 0, (LPARAM) 0);
 				return;
 			} else {
-				sendFileUploadDataMsg(dwNumBytesRead, pBuf);
+				sendFileUploadDataMsg((unsigned short)dwNumBytesRead, pBuf);
 				PostMessage(m_pCC->m_hwnd, (UINT) WM_FT_UPLOADFILEPORTION, (WPARAM) 0, (LPARAM) 0);
 				return;
 			}
@@ -1093,7 +1093,7 @@ FileTransfer::procFLRRename(unsigned short numFiles, FileInfo *pFI)
 int
 FileTransfer::isExistName(FileInfo *pFI, char *pName)
 {
-	for (int i = 0; i < pFI->getNumEntries(); i++) {
+	for (unsigned int i = 0; i < pFI->getNumEntries(); i++) {
 		if (strcmp(pFI->getNameAt(i), pName) == 0) {
 			return i;
 		}
